+++ src/Events/Ready.ts
@@ -1,11 +1,22 @@
 import { Event } from "../Structures/Event.js";
-import { ClientEvents } from "discord.js";
+import { ClientEvents, MessageEmbed } from "discord.js";
 import { formatDate } from "../lib/Utility/Date.js";
+import { client } from "../index.js";
+
+import { createRequire } from 'module';
+const req = createRequire(import.meta.url);
+const { botOwner } = req('../../config.json');
 
 export default class implements Event {
     name: keyof ClientEvents = 'ready';
 
-    init() {
-        console.log(`Logged in at ${formatDate('MMMM Do, YYYY hh:mm:ssA', new Date())}`);
+    async init() {
+        const s = `Logged in at ${formatDate('MMMM Do, YYYY hh:mm:ssA', new Date())}`;
+        console.log(s);
+        
+        if(typeof botOwner === 'string') {
+            const user = await client.users.fetch(botOwner);
+            await user.send(new MessageEmbed().setDescription(s).setColor('#ffe449')); 
+        }
     }
 }